requirements.txt
asgiref==3.7.2
asttokens==2.4.1
black==23.11.0
click==8.1.7
decorator==5.1.1
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
executing==2.0.1
getenv==0.2.0
install==1.3.5
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg-binary==3.1.12
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
typing_extensions==4.8.0
wcwidth==0.2.9
